+++ src/main/java/next/controller/HomeController.java
@@ -1,5 +1,6 @@
 package next.controller;
 
+import core.annotation.Inject;
 import core.annotation.web.Controller;
 import core.annotation.web.RequestMapping;
 import core.annotation.web.RequestMethod;
@@ -12,7 +13,12 @@ import javax.servlet.http.HttpServletResponse;
 
 @Controller
 public class HomeController extends AbstractNewController {
-    private QuestionDao questionDao = QuestionDao.getInstance();
+    private final QuestionDao questionDao;
+
+    @Inject
+    public HomeController(QuestionDao questionDao) {
+        this.questionDao = questionDao;
+    }
 
     @RequestMapping(value = "/", method = RequestMethod.GET)
     public ModelAndView execute(HttpServletRequest request, HttpServletResponse response) throws Exception {
