+++ src/main/java/next/controller/QnaController.java
@@ -1,5 +1,6 @@
 package next.controller;
 
+import core.annotation.Inject;
 import core.annotation.web.Controller;
 import core.annotation.web.RequestMapping;
 import core.annotation.web.RequestMethod;
@@ -19,9 +20,16 @@ import java.util.List;
 
 @Controller
 public class QnaController extends AbstractNewController {
-    private QuestionDao questionDao = QuestionDao.getInstance();
-    private AnswerDao answerDao = AnswerDao.getInstance();
-    private QnaService qnaService = QnaService.getInstance();
+    private final QuestionDao questionDao;
+    private final AnswerDao answerDao;
+    private final QnaService qnaService;
+
+    @Inject
+    public QnaController(QuestionDao questionDao, AnswerDao answerDao, QnaService qnaService) {
+        this.questionDao = questionDao;
+        this.answerDao = answerDao;
+        this.qnaService = qnaService;
+    }
 
     @RequestMapping(value = "/qna/form", method = RequestMethod.GET)
     public ModelAndView createForm(HttpServletRequest req, HttpServletResponse resp) throws Exception {
