+++ src/main/java/next/dao/AnswerDao.java
@@ -1,81 +1,16 @@
 package next.dao;
 
-import core.jdbc.JdbcTemplate;
-import core.jdbc.KeyHolder;
-import core.jdbc.PreparedStatementCreator;
-import core.jdbc.RowMapper;
 import next.model.Answer;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
-import java.sql.*;
 import java.util.List;
 
-public class AnswerDao {
-    private static final Logger logger = LoggerFactory.getLogger( AnswerDao.class );
+public interface AnswerDao {
 
-    private static AnswerDao answerDao = new AnswerDao();
-    private JdbcTemplate jdbcTemplate = JdbcTemplate.getInstance();
+    Answer insert(Answer answer);
 
-    private AnswerDao() {
-    }
+    Answer findById(long answerId);
 
-    public static AnswerDao getInstance() {
-        return answerDao;
-    }
+    List<Answer> findAllByQuestionId(long questionId);
 
-    public Answer insert(Answer answer) {
-        String sql = "INSERT INTO ANSWERS (writer, contents, createdDate, questionId) VALUES (?, ?, ?, ?)";
-        PreparedStatementCreator psc = new PreparedStatementCreator() {
-            @Override
-            public PreparedStatement createPreparedStatement(Connection con) throws SQLException {
-                PreparedStatement pstmt = con.prepareStatement(sql);
-                pstmt.setString(1, answer.getWriter());
-                pstmt.setString(2, answer.getContents());
-                pstmt.setTimestamp(3, new Timestamp(answer.getTimeFromCreateDate()));
-                pstmt.setLong(4, answer.getQuestionId());
-                return pstmt;
-            }
-        };
-
-        KeyHolder keyHolder = new KeyHolder();
-        jdbcTemplate.update(psc, keyHolder);
-        logger.debug("KeyHolder : {}", keyHolder);
-        return findById(keyHolder.getId());
-    }
-
-    public Answer findById(long answerId) {
-        logger.debug("find AnswerId : {}", answerId);
-        String sql = "SELECT answerId, writer, contents, createdDate, questionId FROM ANSWERS WHERE answerId = ?";
-
-        RowMapper<Answer> rm = new RowMapper<Answer>() {
-            @Override
-            public Answer mapRow(ResultSet rs) throws SQLException {
-                return new Answer(rs.getLong("answerId"), rs.getString("writer"), rs.getString("contents"),
-                        rs.getTimestamp("createdDate"), rs.getLong("questionId"));
-            }
-        };
-
-        return jdbcTemplate.queryForObject(sql, rm, answerId);
-    }
-
-    public List<Answer> findAllByQuestionId(long questionId) {
-        String sql = "SELECT answerId, writer, contents, createdDate FROM ANSWERS WHERE questionId = ? "
-                + "order by answerId desc";
-
-        RowMapper<Answer> rm = new RowMapper<Answer>() {
-            @Override
-            public Answer mapRow(ResultSet rs) throws SQLException {
-                return new Answer(rs.getLong("answerId"), rs.getString("writer"), rs.getString("contents"),
-                        rs.getTimestamp("createdDate"), questionId);
-            }
-        };
-
-        return jdbcTemplate.query(sql, rm, questionId);
-    }
-
-    public void delete(Long answerId) {
-        String sql = "DELETE FROM ANSWERS WHERE answerId = ?";
-        jdbcTemplate.update(sql, answerId);
-    }
+    void delete(Long answerId);
 }
