+++ src/main/java/next/dao/QuestionDao.java
@@ -1,88 +1,20 @@
 package next.dao;
 
-import core.jdbc.JdbcTemplate;
-import core.jdbc.KeyHolder;
-import core.jdbc.PreparedStatementCreator;
-import core.jdbc.RowMapper;
 import next.model.Question;
 
-import java.sql.*;
 import java.util.List;
 
-public class QuestionDao {
-    private static QuestionDao questionDao = new QuestionDao();
+public interface QuestionDao {
 
-    private JdbcTemplate jdbcTemplate = JdbcTemplate.getInstance();
+    public Question insert(Question question);
 
-    private QuestionDao() {
-    }
+    public List<Question> findAll();
 
-    public static QuestionDao getInstance() {
-        return questionDao;
-    }
+    public Question findById(long questionId);
 
-    public Question insert(Question question) {
-        String sql = "INSERT INTO QUESTIONS (writer, title, contents, createdDate) VALUES (?, ?, ?, ?)";
-        PreparedStatementCreator psc = new PreparedStatementCreator() {
-            @Override
-            public PreparedStatement createPreparedStatement(Connection con) throws SQLException {
-                PreparedStatement pstmt = con.prepareStatement(sql);
-                pstmt.setString(1, question.getWriter());
-                pstmt.setString(2, question.getTitle());
-                pstmt.setString(3, question.getContents());
-                pstmt.setTimestamp(4, new Timestamp(question.getTimeFromCreateDate()));
-                return pstmt;
-            }
-        };
+    public void update(Question question);
 
-        KeyHolder keyHolder = new KeyHolder();
-        jdbcTemplate.update(psc, keyHolder);
-        return findById(keyHolder.getId());
-    }
+    public void delete(long questionId);
 
-    public List<Question> findAll() {
-        String sql = "SELECT questionId, writer, title, createdDate, countOfAnswer FROM QUESTIONS "
-                + "order by questionId desc";
-
-        RowMapper<Question> rm = new RowMapper<Question>() {
-            @Override
-            public Question mapRow(ResultSet rs) throws SQLException {
-                return new Question(rs.getLong("questionId"), rs.getString("writer"), rs.getString("title"), null,
-                        rs.getTimestamp("createdDate"), rs.getInt("countOfAnswer"));
-            }
-
-        };
-
-        return jdbcTemplate.query(sql, rm);
-    }
-
-    public Question findById(long questionId) {
-        String sql = "SELECT questionId, writer, title, contents, createdDate, countOfAnswer FROM QUESTIONS "
-                + "WHERE questionId = ?";
-
-        RowMapper<Question> rm = new RowMapper<Question>() {
-            @Override
-            public Question mapRow(ResultSet rs) throws SQLException {
-                return new Question(rs.getLong("questionId"), rs.getString("writer"), rs.getString("title"),
-                        rs.getString("contents"), rs.getTimestamp("createdDate"), rs.getInt("countOfAnswer"));
-            }
-        };
-
-        return jdbcTemplate.queryForObject(sql, rm, questionId);
-    }
-
-    public void update(Question question) {
-        String sql = "UPDATE QUESTIONS set title = ?, contents = ? WHERE questionId = ?";
-        jdbcTemplate.update(sql, question.getTitle(), question.getContents(), question.getQuestionId());
-    }
-
-    public void delete(long questionId) {
-        String sql = "DELETE FROM QUESTIONS WHERE questionId = ?";
-        jdbcTemplate.update(sql, questionId);
-    }
-
-    public void updateCountOfAnswer(long questionId) {
-        String sql = "UPDATE QUESTIONS set countOfAnswer = countOfAnswer + 1 WHERE questionId = ?";
-        jdbcTemplate.update(sql, questionId);
-    }
+    public void updateCountOfAnswer(long questionId);
 }
